Updates to refactorings
