+++ frontend/src/actions/productActions.js
@@ -1,5 +1,6 @@
 import axios from 'axios';
 import { productsFail, productsRequest, productsSuccess } from '../slices/productsSlice';
+import { productFail, productRequest, productSuccess } from '../slices/productSlice';
 
 export const getProducts = (keyword, price, category, rating, currentPage) => async (dispatch)=>{
     try {
@@ -27,5 +28,16 @@ export const getProducts = (keyword, price, category, rating, currentPage) => as
     } catch (error) {
         //handle error
         dispatch(productsFail(error.response.data.message))
+    }
+}
+
+export const getProduct = id => async (dispatch)=>{
+    try {
+        dispatch(productRequest())
+        const {data} = await axios.get(`/api/v1/product/${id}`)
+        dispatch(productSuccess(data))
+    } catch (error) {
+        //handle error
+        dispatch(productFail(error.response.data.message))
     }
 }
